removing the common folder and renaming the js_logger.hpp to logger.hpp
